requirements.txt
attrs==23.2.0
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
fake-useragent==1.5.1
google-api-core==2.16.2
google-api-python-client>=2.94.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
gspread==6.0.0
h11==0.14.0
httplib2==0.22.0
idna==3.6
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==23.2
pillow==10.2.0
protobuf==4.25.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.1
PySocks==1.7.1
python-dotenv==1.0.1
retry
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
selenium==4.24.0
sniffio==1.3.0
sortedcontainers==2.4.0
StrEnum==0.4.15
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
undetected-chromedriver==3.5.5
uritemplate==4.1.1
urllib3==2.2.0
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==13.0.1
wsproto==1.2.0
